+++ EmmaSharp/Models/Members/Member.cs
@@ -22,7 +22,7 @@ namespace EmmaSharp.Models.Members
         public Dictionary<string, string> Fields { get; set; }
 
         [JsonProperty("member_id")]
-        public long? MemberId { get; set; }
+        public int? MemberId { get; set; }
 
         [JsonConverter(typeof(EmmaDateConverter))]
         [JsonProperty("last_modified_at")]
